Refactor application for scalibility
