+++ middleware/auth_middleware.js
@@ -1,14 +1,16 @@
 const jwt = require("jsonwebtoken");
+require("dotenv").config();
 
-exports.cookieJwtAuth = (req, res, next) => {
-  const token = req.cookies.token;
+exports.validate = (req, res, next) => {
+  const token = req.get("token");
   try {
     const user = jwt.verify(token, process.env.AUTH_SECRET);
     req.user = user;
-    console.log("user = ", user);
+    req.email = user.email;
     next();
   } catch (err) {
     console.log(err);
+    console.log("clearing the token");
     res.clearCookie("token").end();
     throw new Error("Access denied exception. Token is expired.");
   }
