+++ middleware/time_middleware.js
@@ -1,0 +1,7 @@
+"use strict";
+
+exports.requestTime = function (req, res, next) {
+  req.requestTime = Date.now();
+  console.log("Time:", req.requestTime);
+  next();
+};
