+++ package.json
@@ -3,8 +3,9 @@
     "cookie-parser": "^1.4.6",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "jsonwebtoken": "^8.5.1",
+    "jsonwebtoken": "^9.0.0",
     "mongodb": "^4.12.1",
+    "nodemailer": "^6.8.0",
     "nodemon": "^2.0.20"
   },
   "engines": {
